sistemas-de-recomendacao/tp2/requirements.txt
numpy==1.23.5
scikit-learn==1.1.3
pandas==1.5.1
scikit-surprise==1.1.3
